+++ src/layman/authz/read_everyone_write_owner.py
@@ -43,3 +43,7 @@ def get_gs_roles(username, type):
     elif type == 'w':
         roles = gs.get_roles_owner(username)
     return roles
+
+
+def can_i_edit(publ_type, username, publication_name):
+    return g.user is not None and g.user['username'] == username